requirements.txt
python-dotenv==1.0.1
langchain==0.1.3
langchain_core==0.1.15
langchain_openai==0.0.3
flask==3.0.2
flask_cors==4.0.0
requests==2.31.0
